# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Utilisateur API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
...
        "/utilisateurs": {
            "post": {
                "summary": "Créer un utilisateur",
                "description": "Crée un nouvel utilisateur dans le système.",
                "operationId": "90475ae6f2f66546b4d34fde91ffcca5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "nom",
                                    "prenom",
                                    "date_naissance",
                                    "mot_de_passe",
                                    "etat",
                                    "nb_tentative",
                                    "id_sexe"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "nom": {
                                        "type": "string",
                                        "example": "Dupont"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Jean"
                                    },
                                    "date_naissance": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "mot_de_passe": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "etat": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "nb_tentative": {
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "id_sexe": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "email": {
                                            "type": "string",
                                            "example": "user@example.com"
                                        },
                                        "nom": {
                                            "type": "string",
                                            "example": "Dupont"
                                        },
                                        "prenom": {
                                            "type": "string",
                                            "example": "Jean"
                                        },
                                        "date_naissance": {
                                            "type": "string",
                                            "example": "1990-01-01"
                                        },
                                        "mot_de_passe": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "etat": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "nb_tentative": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "id_sexe": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/utilisateurs/{id}": {
            "put": {
                "summary": "Mettre à jour un utilisateur",
                "description": "Met à jour un utilisateur avec de nouvelles informations.",
                "operationId": "a8b653d914618ceb46ea741708d16fd1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Dupont"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Jean"
                                    },
                                    "date_naissance": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "mot_de_passe": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "id_sexe": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "email": {
                                            "type": "string",
                                            "example": "user@example.com"
                                        },
                                        "nom": {
                                            "type": "string",
                                            "example": "Dupont"
                                        },
                                        "prenom": {
                                            "type": "string",
                                            "example": "Jean"
                                        },
                                        "date_naissance": {
                                            "type": "string",
                                            "example": "1990-01-01"
                                        },
                                        "mot_de_passe": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "etat": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "nb_tentative": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "id_sexe": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Session expirée"
                    }
                }
            }
        },
        "/utilisateurs/inscrire-utilisateur": {
            "post": {
                "summary": "Inscrire un utilisateur",
                "description": "Inscrit un utilisateur dans le système en envoyant un lien de validation.",
                "operationId": "9dd9be9d6181efbaa6bce0fe9544b1c8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "nom",
                                    "prenom",
                                    "date_naissance",
                                    "mot_de_passe",
                                    "sexe"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "nom": {
                                        "type": "string",
                                        "example": "Dupont"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Jean"
                                    },
                                    "date_naissance": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "mot_de_passe": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "sexe": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur inscrit avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "url": {
                                            "type": "string",
                                            "example": "http://example.com/validate"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
...
    },
...
}